+++ tests/checks/scaling_phase_custom_resource/scaling_phase_custom_resource_test.go
@@ -1,41 +1,38 @@
 //go:build e2e
 // +build e2e
 
-package ingress_in_keda_namespace_test
+package scaling_phase_test
 
 import (
 	"fmt"
+	"strconv"
+	"strings"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
-	"k8s.io/client-go/kubernetes"
 
 	. "github.com/kedacore/http-add-on/tests/helper"
 )
 
 const (
-	testName = "ingress-in-keda-namespace-test"
+	testName = "scaling-phase-custom-resource-test"
 )
 
 var (
 	testNamespace        = fmt.Sprintf("%s-ns", testName)
-	deploymentName       = fmt.Sprintf("%s-deployment", testName)
+	rolloutName          = fmt.Sprintf("%s-rollout", testName)
 	serviceName          = fmt.Sprintf("%s-service", testName)
-	ingressName          = fmt.Sprintf("%s-ingress", testName)
 	httpScaledObjectName = fmt.Sprintf("%s-http-so", testName)
-	ingressHost          = fmt.Sprintf("http://%s-controller.%s", IngressReleaseName, IngressNamespace)
 	host                 = testName
 	minReplicaCount      = 0
-	maxReplicaCount      = 1
+	maxReplicaCount      = 4
 )
 
 type templateData struct {
 	TestNamespace        string
-	DeploymentName       string
+	RolloutName          string
 	ServiceName          string
-	IngressName          string
-	KEDANamespace        string
-	IngressHost          string
 	HTTPScaledObjectName string
 	Host                 string
 	MinReplicas          int
@@ -43,29 +40,6 @@ type templateData struct {
 }
 
 const (
-	ingressTemplate = `
-apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: {{.IngressName}}
-  namespace: {{.KEDANamespace}}
-  annotations:
-    nginx.ingress.kubernetes.io/rewrite-target: /
-    kubernetes.io/ingress.class: nginx
-spec:
-  rules:
-  - host: {{.Host}}
-    http:
-      paths:
-      - path: /
-        pathType: Prefix
-        backend:
-          service:
-            name: keda-http-add-on-interceptor-proxy
-            port:
-              number: 8080
-`
-
 	serviceTemplate = `
 apiVersion: v1
 kind: Service
@@ -73,7 +47,7 @@ metadata:
   name: {{.ServiceName}}
   namespace: {{.TestNamespace}}
   labels:
-    app: {{.DeploymentName}}
+    app: {{.RolloutName}}
 spec:
   ports:
     - port: 8080
@@ -81,29 +55,34 @@ spec:
       protocol: TCP
       name: http
   selector:
-    app: {{.DeploymentName}}
+    app: {{.RolloutName}}
 `
 
-	deploymentTemplate = `
-apiVersion: apps/v1
-kind: Deployment
+	workloadTemplate = `
+apiVersion: argoproj.io/v1alpha1
+kind: Rollout
 metadata:
-  name: {{.DeploymentName}}
+  name: {{.RolloutName}}
   namespace: {{.TestNamespace}}
   labels:
-    app: {{.DeploymentName}}
+    app: {{.RolloutName}}
 spec:
   replicas: 0
+  strategy:
+    canary:
+      steps:
+        - setWeight: 50
+        - pause: {duration: 1}
   selector:
     matchLabels:
-      app: {{.DeploymentName}}
+      app: {{.RolloutName}}
   template:
     metadata:
       labels:
-        app: {{.DeploymentName}}
+        app: {{.RolloutName}}
     spec:
       containers:
-        - name: {{.DeploymentName}}
+        - name: {{.RolloutName}}
           image: registry.k8s.io/e2e-test-images/agnhost:2.45
           args:
           - netexec
@@ -121,21 +100,28 @@ spec:
 apiVersion: batch/v1
 kind: Job
 metadata:
-  name: generate-request
+  name: load-generator
   namespace: {{.TestNamespace}}
 spec:
   template:
     spec:
       containers:
-      - name: curl-client
-        image: curlimages/curl
+      - name: apache-ab
+        image: ghcr.io/kedacore/tests-apache-ab
         imagePullPolicy: Always
-        command: ["curl", "-H", "Host: {{.Host}}", "{{.IngressHost}}"]
+        args:
+          - "-n"
+          - "2000000"
+          - "-c"
+          - "20"
+          - "-H"
+          - "Host: {{.Host}}"
+          - "http://keda-http-add-on-interceptor-proxy.keda:8080/"
       restartPolicy: Never
+      terminationGracePeriodSeconds: 5
   activeDeadlineSeconds: 600
   backoffLimit: 5
 `
-
 	httpScaledObjectTemplate = `
 kind: HTTPScaledObject
 apiVersion: http.keda.sh/v1alpha1
@@ -145,10 +131,12 @@ metadata:
 spec:
   hosts:
   - {{.Host}}
-  targetPendingRequests: 100
+  targetPendingRequests: 2
   scaledownPeriod: 10
   scaleTargetRef:
-    deployment: {{.DeploymentName}}
+    name: {{.RolloutName}}
+    apiVersion: argoproj.io/v1alpha1
+    kind: Rollout
     service: {{.ServiceName}}
     port: 8080
   replicas:
@@ -165,51 +153,70 @@ func TestCheck(t *testing.T) {
 	data, templates := getTemplateData()
 	CreateKubernetesResources(t, kc, testNamespace, data, templates)
 
-	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, minReplicaCount, 6, 10),
+	assert.True(t, waitForArgoRolloutReplicaCount(t, rolloutName, testNamespace, minReplicaCount, 6, 10),
 		"replica count should be %d after 1 minutes", minReplicaCount)
-	assert.True(t, WaitForIngressReady(t, kc, ingressName, KEDANamespace, 12, 10),
-		"ingress should be ready after 2 minutes")
 
-	testScaleOut(t, kc, data)
-	testScaleIn(t, kc, data)
+	testScaleOut(t, data)
+	testScaleIn(t)
 
 	// cleanup
 	DeleteKubernetesResources(t, testNamespace, data, templates)
 }
 
-func testScaleOut(t *testing.T, kc *kubernetes.Clientset, data templateData) {
+func testScaleOut(t *testing.T, data templateData) {
 	t.Log("--- testing scale out ---")
 
 	KubectlApplyWithTemplate(t, data, "loadJobTemplate", loadJobTemplate)
 
-	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, maxReplicaCount, 6, 10),
-		"replica count should be %d after 1 minutes", maxReplicaCount)
+	assert.True(t, waitForArgoRolloutReplicaCount(t, rolloutName, testNamespace, maxReplicaCount, 18, 10),
+		"replica count should be %d after 3 minutes", maxReplicaCount)
+	KubectlDeleteWithTemplate(t, data, "loadJobTemplate", loadJobTemplate)
 }
 
-func testScaleIn(t *testing.T, kc *kubernetes.Clientset, data templateData) {
+func testScaleIn(t *testing.T) {
 	t.Log("--- testing scale out ---")
 
-	KubectlDeleteWithTemplate(t, data, "loadJobTemplate", loadJobTemplate)
-	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, minReplicaCount, 12, 10),
+	assert.True(t, waitForArgoRolloutReplicaCount(t, rolloutName, testNamespace, minReplicaCount, 12, 10),
 		"replica count should be %d after 2 minutes", minReplicaCount)
 }
 
 func getTemplateData() (templateData, []Template) {
 	return templateData{
 			TestNamespace:        testNamespace,
-			DeploymentName:       deploymentName,
+			RolloutName:          rolloutName,
 			ServiceName:          serviceName,
-			IngressName:          ingressName,
-			KEDANamespace:        KEDANamespace,
 			HTTPScaledObjectName: httpScaledObjectName,
-			IngressHost:          ingressHost,
 			Host:                 host,
 			MinReplicas:          minReplicaCount,
 			MaxReplicas:          maxReplicaCount,
 		}, []Template{
-			{Name: "deploymentTemplate", Config: deploymentTemplate},
-			{Name: "serviceTemplate", Config: serviceTemplate},
-			{Name: "ingressTemplate", Config: ingressTemplate},
+			{Name: "workloadTemplate", Config: workloadTemplate},
+			{Name: "serviceNameTemplate", Config: serviceTemplate},
 			{Name: "httpScaledObjectTemplate", Config: httpScaledObjectTemplate},
 		}
+}
+
+func waitForArgoRolloutReplicaCount(t *testing.T, name, namespace string,
+	target, iterations, intervalSeconds int) bool {
+	for i := 0; i < iterations; i++ {
+		kctlGetCmd := fmt.Sprintf(`kubectl get rollouts.argoproj.io/%s -n %s -o jsonpath="{.spec.replicas}"`, name, namespace)
+		output, err := ExecuteCommand(kctlGetCmd)
+
+		assert.NoErrorf(t, err, "cannot get rollout info - %s", err)
+
+		unqoutedOutput := strings.ReplaceAll(string(output), "\"", "")
+		replicas, err := strconv.ParseInt(unqoutedOutput, 10, 64)
+		assert.NoErrorf(t, err, "cannot convert rollout count to int - %s", err)
+
+		t.Logf("Waiting for rollout replicas to hit target. Deployment - %s, Current  - %d, Target - %d",
+			name, replicas, target)
+
+		if replicas == int64(target) {
+			return true
+		}
+
+		time.Sleep(time.Duration(intervalSeconds) * time.Second)
+	}
+
+	return false
 }